refactor: Rename AddContentContainer to AddContent
to avoid confusion with library containers. The AddContent is used to add content to libraries, collections and containers.
