+++ src/library-authoring/LibraryLayout.tsx
@@ -43,7 +43,7 @@ const LibraryLayout = () => {
       /** The component picker modal to use. We need to pass it as a reference instead of
        * directly importing it to avoid the import cycle:
        * ComponentPicker > LibraryAuthoringPage/LibraryCollectionPage >
-       * Sidebar > AddContentContainer > ComponentPicker */
+       * Sidebar > AddContent > ComponentPicker */
       componentPicker={ComponentPicker}
     >
       <SidebarProvider>
