+++ src/library-authoring/add-content/index.ts
@@ -1,2 +1,2 @@
-export { default as AddContentContainer } from './AddContentContainer';
+export { default as AddContent } from './AddContent';
 export { default as AddContentHeader } from './AddContentHeader';
